chore: 로그인이 안되어 있을 경우 로그인으로 보내는 코드를 옮깁니다
민호님께서 작성해주신 해당 로직을 ProtectedRoute 안으로 옮겼습니다.
